🔧 Enforce `server-only` Usage for App Config and Client App Config * Refactors `appConfig.ts` and `clientConfig.ts` to support `server-only` usage, splitting type definitions out into separate `types.ts` in `src/config/` * Adds import for `server-only` package on `appConfig.ts` and `clientConfig.ts`
