+++ apps/consent-ui/src/config/clientConfig.ts
@@ -17,30 +17,9 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { defaultAppConfig } from 'src/config';
+import 'server-only';
 
-/**
- * Environment variables exposed to client components via AppConfigContext
- *
- * Note: this config is exposed to the client, so it should not contain any sensitive information
- */
-export type ClientAppConfig = {
-	CONSENT_API_URL: string;
-	CONSENT_UI_URL: string;
-	FEATURE_CONSENT_PDF_BUTTONS?: boolean;
-	OHCRN_EMAIL?: string;
-	OHCRN_HOME_LINK?: string;
-	RECAPTCHA_SITE_KEY?: string;
-};
-
-export const defaultClientAppConfig: ClientAppConfig = {
-	CONSENT_API_URL: defaultAppConfig.CONSENT_API_URL,
-	CONSENT_UI_URL: defaultAppConfig.CONSENT_UI_URL,
-	FEATURE_CONSENT_PDF_BUTTONS: defaultAppConfig.FEATURE_CONSENT_PDF_BUTTONS,
-	OHCRN_EMAIL: defaultAppConfig.OHCRN_EMAIL,
-	OHCRN_HOME_LINK: defaultAppConfig.OHCRN_HOME_LINK,
-	RECAPTCHA_SITE_KEY: defaultAppConfig.RECAPTCHA_SITE_KEY,
-};
+import { defaultAppConfig, ClientAppConfig } from 'src/config/types';
 
 /**
  * Returns environment variables for client components
@@ -48,7 +27,6 @@ export const defaultClientAppConfig: ClientAppConfig = {
  * Note: intended for **SERVER COMPONENTS ONLY**. Access these values from the client using the `useAppConfigContext` hook.
  * @returns {ClientAppConfig}
  */
-// TODO: enforce server-only usage. To be completed as a follow-up to https://github.com/OHCRN/platform/issues/422
 const getClientAppConfig = (): ClientAppConfig => ({
 	CONSENT_API_URL: process.env.CONSENT_API_URL || defaultAppConfig.CONSENT_API_URL,
 	CONSENT_UI_URL: process.env.CONSENT_UI_URL || defaultAppConfig.CONSENT_UI_URL,
